Organize files into packages
