+++ src/main/java/duke/Main.java
@@ -1,6 +1,6 @@
 package duke;
 
-import duke.controllers.MainWindow;
+import duke.controller.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -10,6 +10,9 @@ import javafx.stage.Stage;
 import java.io.IOException;
 
 
+/**
+ * GUI for Duke
+ */
 public class Main extends Application {
 
     private Duke duke = new Duke("./data");
