+++ src/main/java/duke/Storage.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.exception.DukeException;
 import duke.task.TaskList;
 
 import java.io.File;
