+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 import duke.task.Deadline;
 import duke.task.TaskList;
