+++ src/main/java/duke/command/FindCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 import duke.task.TaskList;
 
 public class FindCommand extends Command {
