+++ src/main/java/duke/command/ListCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 import duke.Ui;
 import duke.task.TaskList;
