+++ src/main/java/duke/command/TodoCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 import duke.task.TaskList;
 import duke.task.Todo;
