+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 import duke.task.TaskList;
 
 public class UnmarkCommand extends Command {
