+++ src/main/java/duke/controller/DialogBox.java
@@ -1,4 +1,4 @@
-package duke.controllers;
+package duke.controller;
 
 import java.io.IOException;
 import java.util.Collections;
