+++ src/main/java/duke/task/Deadline.java
@@ -2,9 +2,6 @@ package duke.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-
-import duke.DukeException;
 
 public class Deadline extends Task {
     private LocalDate by;
