+++ src/main/java/duke/task/Event.java
@@ -2,9 +2,8 @@ package duke.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 
 public class Event extends Task {
     private LocalDate at;
