+++ src/main/resources/view/DialogBox.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="-Infinity" prefHeight="108.0" prefWidth="392.0" style="-fx-background-color: #21130d; -fx-background-radius: 20px; -fx-opacity: 90%;" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1">
+<fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="-Infinity" prefHeight="108.0" prefWidth="392.0" style="-fx-background-color: #21130d; -fx-background-radius: 20px; -fx-opacity: 90%;" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
     <children>
       <TextFlow maxHeight="1.7976931348623157E308" maxWidth="-Infinity">
          <HBox.margin>
